- Get the dashboard data from the database.
- User can log in using an admin account.
- User can log out.
- Disable editing features when the user is logged out.
- Add Guild Master styling to the authentication page.

+++ src/controllers/handlers/faction/FactionRD.java
@@ -9,6 +9,7 @@ import jakarta.servlet.http.HttpServletResponse;
 import jakarta.servlet.ServletException;
 
 import models.Faction;
+import utils.AuthenticationSecurity;
 import utils.ExceptionHandler;
 import utils.RequestChecker;
 
@@ -16,7 +17,7 @@ public class FactionRD extends HttpServlet {
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
         try {
-            if (RequestChecker.isDeleteMode(req)) {
+            if (RequestChecker.isDeleteMode(req) && AuthenticationSecurity.isLoggedIn(req)) {
                 int factionId = Integer.parseInt(req.getParameter("faction-id"));
                 Faction.deleteById(factionId);
             }
